requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-multipart==0.0.20
yt-dlp==2025.1.26
opencv-python==4.11.0.86
pillow==11.1.0
pydantic==2.10.6
streamlit>=1.41.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.37
alembic>=1.14.1
pytest>=8.3.4
httpx>=0.28.1
python-dotenv>=1.0.1
pydantic-settings>=2.7.1
openai==1.61.0